Move interfaces to own namespace
