+++ LazyStorage/InMemory/InMemoryRepositoryWithConverter.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using LazyStorage.Interfaces;
 
 namespace LazyStorage.InMemory
 {
