+++ LazyStorage/InMemory/InMemoryStorage.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using LazyStorage.Interfaces;
 
 namespace LazyStorage.InMemory
 {
