+++ LazyStorage/Interfaces/IStorage.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace LazyStorage
+﻿namespace LazyStorage.Interfaces
 {
     public interface IStorage
     {
