+++ LazyStorage/LazyStorage.csproj
@@ -41,12 +41,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="IConverter.cs" />
+    <Compile Include="Interfaces\IConverter.cs" />
     <Compile Include="InMemory\InMemoryRepositoryWithConverter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="IRepository.cs" />
-    <Compile Include="IStorable.cs" />
-    <Compile Include="IStorage.cs" />
+    <Compile Include="Interfaces\IRepository.cs" />
+    <Compile Include="Interfaces\IStorable.cs" />
+    <Compile Include="Interfaces\IStorage.cs" />
     <Compile Include="InMemory\InMemoryRepository.cs" />
     <Compile Include="InMemory\InMemorySingleton.cs" />
     <Compile Include="InMemory\InMemoryStorage.cs" />
