+++ LazyStorage/Xml/XmlRepository.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.Serialization;
 using System.Xml.Linq;
+using LazyStorage.Interfaces;
 
 namespace LazyStorage.Xml
 {
