+++ LazyStorage/Xml/XmlStorage.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Xml.Linq;
+using LazyStorage.Interfaces;
 
 namespace LazyStorage.Xml
 {
