+++ UnitTests/RepositoryTests.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Xml.Linq;
 using LazyStorage.InMemory;
+using LazyStorage.Interfaces;
 using LazyStorage.Xml;
 using Xunit;
