+++ UnitTests/TestObjectNotIStorable.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Runtime.Serialization;
+using LazyStorage.Interfaces;
 
 namespace LazyStorage.Tests
 {
